requirements.txt
anthropic
openai
google-generativeai
tqdm
config
argparse
inflect
